FontManager implementation
SFML 2.1 does not load a default font by default so it becomes mandatory to
load at least one font and use it when using a Text object, FontManager is a
utility class to load and manage fonts. It will search and load fonts in
directories provided by the user so that it is possible for the user to load
fonts by simply providing a font name (e.g. "arial.ttf")